ReActXen/src/reactxen/demo/intent_implementation_demo/multi_agent_implementation_demo/performance_optimizer.py
"""
Performance Optimizer with Multithreading and Hardware Acceleration
Optimizes agent execution for speed and efficiency.
"""
import threading
import queue
import time
from typing import Dict, List, Any, Optional, Callable
from concurrent.futures import ThreadPoolExecutor, as_completed
from pathlib import Path


class PerformanceOptimizer:
    """Optimizes agent performance with multithreading and hardware acceleration."""
    
    def __init__(self, max_workers: int = 4):
        self.max_workers = max_workers
        self.executor = ThreadPoolExecutor(max_workers=max_workers)
        self.task_queue = queue.Queue()
        self.results = {}
        
    def parallel_tool_execution(
        self,
        tools_and_inputs: List[tuple[Callable, Dict[str, Any]]],
        timeout: float = 30.0
    ) -> Dict[str, Any]:
        """Execute multiple tools in parallel."""
        results = {}
        
        def execute_tool(tool_func, inputs):
            try:
                start = time.time()
                result = tool_func(**inputs)
                execution_time = time.time() - start
                return {
                    "success": True,
                    "result": result,
                    "execution_time": execution_time
                }
            except Exception as e:
                return {
                    "success": False,
                    "error": str(e),
                    "execution_time": 0.0
                }
        
        # Submit all tasks
        futures = {}
        for i, (tool_func, inputs) in enumerate(tools_and_inputs):
            future = self.executor.submit(execute_tool, tool_func, inputs)
            futures[future] = f"tool_{i}"
        
        # Collect results
        for future in as_completed(futures, timeout=timeout):
            tool_id = futures[future]
            try:
                results[tool_id] = future.result(timeout=1.0)
            except Exception as e:
                results[tool_id] = {
                    "success": False,
                    "error": f"Timeout or error: {str(e)}",
                    "execution_time": timeout
                }
        
        return results
    
    def batch_dataset_operations(
        self,
        datasets: List[str],
        operation: Callable,
        **kwargs
    ) -> Dict[str, Any]:
        """Execute dataset operations in parallel."""
        tasks = [(operation, {"dataset_name": ds, **kwargs}) for ds in datasets]
        return self.parallel_tool_execution(tasks)
    
    def optimized_reflection(self, reasoning_chains: List[Dict], reflector: Callable) -> List[Dict]:
        """Run reflection on multiple reasoning chains in parallel."""
        tasks = [(reflector, {"reasoning_chain": chain}) for chain in reasoning_chains]
        results = self.parallel_tool_execution(tasks, timeout=10.0)
        
        reflections = []
        for tool_id, result in results.items():
            if result["success"]:
                reflections.append({
                    "chain_id": tool_id,
                    "reflection": result["result"],
                    "execution_time": result["execution_time"]
                })
        
        return reflections
    
    def shutdown(self):
        """Shutdown the executor."""
        self.executor.shutdown(wait=True)


# macOS Metal Performance Shaders integration (if available)
try:
    import Metal
    METAL_AVAILABLE = True
except ImportError:
    METAL_AVAILABLE = False


class MetalAccelerator:
    """Hardware acceleration using macOS Metal Performance Shaders."""
    
    def __init__(self):
        self.available = METAL_AVAILABLE
        if self.available:
            try:
                self.device = Metal.MTLCreateSystemDefaultDevice()
                self.command_queue = self.device.newCommandQueue()
            except:
                self.available = False
    
    def accelerate_computation(self, data: Any, operation: str) -> Any:
        """Accelerate computation using Metal if available."""
        if not self.available:
            return data  # Fallback to CPU
        
        # Metal acceleration would go here for specific operations
        # For now, just return data (placeholder for future implementation)
        return data
    
    def is_available(self) -> bool:
        """Check if Metal acceleration is available."""
        return self.available


def get_performance_optimizer(max_workers: int = 4) -> PerformanceOptimizer:
    """Get or create performance optimizer instance."""
    if not hasattr(get_performance_optimizer, '_instance'):
        get_performance_optimizer._instance = PerformanceOptimizer(max_workers=max_workers)
    return get_performance_optimizer._instance


def get_metal_accelerator() -> MetalAccelerator:
    """Get or create Metal accelerator instance."""
    if not hasattr(get_metal_accelerator, '_instance'):
        get_metal_accelerator._instance = MetalAccelerator()
    return get_metal_accelerator._instance

