ReActXen/src/reactxen/demo/intent_implementation_demo/pyproject.toml
[project]
name = "intent-implementation-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "colorama>=0.4.6",
    "datasets>=4.4.1",
    "ddgs>=9.9.0",
    "haversine>=2.9.0",
    "huggingface-hub>=0.36.0",
    "ibm-watsonx-ai>=1.3.42",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "langchain-ibm>=1.0.0",
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "openai>=1.0.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "sentencepiece>=0.2.1",
    "tiktoken>=0.12.0",
    "mdextractor>=2025.4.0",
    "torch>=2.9.0",
    "torchaudio>=2.9.0",
    "torchvision>=0.24.0",
    "transformers>=4.57.1",
    "xgboost>=3.1.1",
    "tabulate>=0.9.0",
]
